setup authentication

+++ controllers/teas.py
@@ -1,9 +1,11 @@
 from fastapi import APIRouter, HTTPException, Depends
 from sqlalchemy.orm import Session
 from models.tea import TeaModel
-from serializers.tea import TeaSchema
+from models.user import UserModel
+from serializers.tea import TeaSchema, TeaCreateSchema
 from typing import List
 from database import get_db
+from dependencies.get_current_user import get_current_user
 
 router = APIRouter()
 
@@ -21,8 +23,8 @@ def get_single_tea(tea_id: int, db: Session = Depends(get_db)):
     raise HTTPException(status_code=404, detail="Tea not found")
 
 @router.post('/teas', response_model=TeaSchema)
-def create_tea(tea: TeaSchema, db: Session = Depends(get_db)):
-  new_tea =  TeaModel(**tea.dict())
+def create_tea(tea: TeaCreateSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
+  new_tea =  TeaModel(**tea.dict(), user_id=current_user.id)
   db.add(new_tea)
   db.commit()
   db.refresh(new_tea)
@@ -31,11 +33,14 @@ def create_tea(tea: TeaSchema, db: Session = Depends(get_db)):
 # teas.py
 
 @router.put("/teas/{tea_id}", response_model=TeaSchema)
-def update_tea(tea_id: int, tea: TeaSchema, db: Session = Depends(get_db)):
+def update_tea(tea_id: int, tea: TeaCreateSchema, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
   db_tea = db.query(TeaModel).filter(TeaModel.id == tea_id).first()
   # If tea was not found, raise an error
   if not db_tea:
     raise HTTPException(status_code=404, detail="Tea not found")
+
+  if db_tea.user_id != current_user.id:
+    raise HTTPException(status_code=403, detail="Operation forbidden")
 
   tea_data = tea.dict(exclude_unset=True)
   for key, value in tea_data.items():
@@ -47,11 +52,14 @@ def update_tea(tea_id: int, tea: TeaSchema, db: Session = Depends(get_db)):
 
 
 @router.delete("/teas/{tea_id}")
-def delete_tea(tea_id: int, db: Session = Depends(get_db)):
+def delete_tea(tea_id: int, db: Session = Depends(get_db), current_user: UserModel = Depends(get_current_user)):
     db_tea = db.query(TeaModel).filter(TeaModel.id == tea_id).first()
     if not db_tea:
       # If tea was not found, raise an error
       raise HTTPException(status_code=404, detail="Tea not found")
+
+    if db_tea.user_id != current_user.id:
+      raise HTTPException(status_code=403, detail="Operation forbidden")
 
     db.delete(db_tea)  # Remove from database
     db.commit()  # Save changes
